Add files via upload
Working python program with bell scheduling system. The bell system had problems at the school that I work at. So, I wrote this Python program and set up a laptop cabled into our PA system's amp to solve the issue and create a working bell system. This is release version 1 and I plan on adding to the program. But, for now, it works. If you want to change the schedules, you'll need to hard code them and change them. In future releases, I plan on adding the functionality to modify the schedules and customize them. I also plan on adding features to adjust which days the schedule runs.